23/09/2024

+++ src/main/java/org/openjfx/precificacao/models/LancamentoCF.java
@@ -15,7 +15,7 @@ public class LancamentoCF {
 
     private int id;
     private int idProjeto;
-    private Float desconto;
+    private double desconto;
     private LocalDate data;
